feat: only return active accommodations
